Fix regular expression escaping
When passing the regular expression to Python, we escape the `\` char in
the Ruby string so that Python sees `'\A__\w+__\Z'`, but this raw string
produce a warning when running in development mode.

We need to tell Python to consider this as a raw string so that it does
not complain.

This fix the warning:
```
<string>:37: DeprecationWarning: invalid escape sequence \A
```